Fix default gui_client assignment
There was a typo and it said `ni` rather than `nil`. Luckily (or unluckily, it
depends on how you see it), the variable `ni` isn't defined anywhere so it
default to... `nil`.

Not worth making another release with this change only as nothing actually
breaks, so it will end up in the next release.